InitContext becomes optional
