+++ package.json
@@ -27,7 +27,7 @@
   "devDependencies": {
     "@eslint/js": "^9.22.0",
     "@fastify/express": "^4.0.1",
-    "@nestjs/common": "^10.4.7",
+    "@nestjs/common": "^11.0.12",
     "@prisma/client": "^5.3.1",
     "@types/co-body": "^6.1.3",
     "@types/dockerode": "^3.3.34",
